LayoutTests/fast/selectors/nth-child-of-basics-2.html

<!doctype html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<style>
#test-root * {
    background-color: red;
}
</style>
<style id="style">
</style>
</head>
<body>
    <div style="display:none" id="test-root">
        <testcaseA id="testcase1" attribute1="value1" class="bar baz bazoo" attribute3="value3">Not empty</testcaseA>
        <testcaseB id="testcase2" class="bar baz bazoo" attribute2="value2" attribute3="value3"></testcaseB>
        <testcaseA id="testcase3" attribute1="value1" class="foo baz bazoo">Not empty</testcaseA>
        <testcaseB id="testcase4" attribute1="value1" class="foo bar bazoo">Not empty</testcaseB>
        <testcaseA id="testcase5" class="foo baz bazoo" attribute2="value2" attribute3="value3">Not empty</testcaseA>
        <testcaseB id="testcase6" attribute1="value1" class="foo bar baz" attribute3="value3">Not empty</testcaseB>
        <testcaseA id="testcase7" attribute1="value1" class="bar baz bazoo">Not empty</testcaseA>
        <testcaseB id="testcase8" attribute1="value1" class="foo baz bazoo">Not empty</testcaseB>
        <testcaseA id="testcase9" class="foo bar bazoo" attribute2="value2" attribute3="value3"></testcaseA>
        <testcaseB id="testcase10" attribute1="value1" class="foo bar bazoo" attribute3="value3">Not empty</testcaseB>
        <testcaseA id="testcase11" attribute1="value1" class="foo bar baz">Not empty</testcaseA>
        <testcaseB id="testcase12" attribute1="value1" class="bar baz bazoo">Not empty</testcaseB>
        <testcaseA id="testcase13" attribute1="value1" class="foo baz bazoo" attribute3="value3">Not empty</testcaseA>
        <testcaseB id="testcase14" class="foo bar baz" attribute2="value2">Not empty</testcaseB>
        <testcaseA id="testcase15" attribute1="value1" class="foo bar bazoo">Not empty</testcaseA>
        <testcaseB id="testcase16" attribute1="value1" class="foo bar baz" attribute3="value3">Not empty</testcaseB>
        <testcaseA id="testcase17" attribute1="value1" class="bar baz bazoo" attribute3="value3">Not empty</testcaseA>
        <testcaseB id="testcase18" attribute1="value1" class="foo baz bazoo">Not empty</testcaseB>
        <testcaseA id="testcase19" attribute1="value1" class="foo bar bazoo">Not empty</testcaseA>
        <testcaseB id="testcase20" class="bar baz bazoo" attribute2="value2" attribute3="value3">Not empty</testcaseB>
        <testcaseA id="testcase21" attribute1="value1" class="foo bar baz" attribute3="value3"></testcaseA>
        <testcaseB id="testcase22" attribute1="value1" class="bar baz bazoo">Not empty</testcaseB>
        <testcaseA id="testcase23" attribute1="value1" class="foo baz bazoo">Not empty</testcaseA>
        <testcaseB id="testcase24" attribute1="value1" class="foo bar bazoo" attribute3="value3">Not empty</testcaseB>
        <testcaseA id="testcase25" attribute1="value1" class="foo bar baz" attribute3="value3">Not empty</testcaseA>
        <testcaseB id="testcase26" attribute1="value1" class="bar baz bazoo">Not empty</testcaseB>
        <testcaseA id="testcase27" class="foo baz bazoo" attribute2="value2"></testcaseA>
        <testcaseB id="testcase28" attribute1="value1" class="foo baz bazoo" attribute3="value3">Not empty</testcaseB>
        <testcaseA id="testcase29" attribute1="value1" class="foo bar bazoo">Not empty</testcaseA>
        <testcaseB id="testcase30" attribute1="value1" class="foo bar baz">Not empty</testcaseB>
    </div>
</body>
<script>
description('Check the basic features of the ":nth-child(An+B of selectorList)" pseudo class.');

function testQuerySelector(selector, expectedIds) {
    shouldBe("document.querySelectorAll('" + selector + "').length", '' + expectedIds.length);
    for (var i = 0; i < expectedIds.length; ++i)
        shouldBeEqualToString("document.querySelectorAll('" + selector + "')[" + i + "].id", expectedIds[i]);
}

function testStyling(selector, expectedIds) {
    var stylingElement = document.getElementById("style");
    stylingElement.innerHTML = '' + selector + ' { background-color: rgb(10, 100, 200); }';

    var allTestCases = document.querySelectorAll("#test-root *");
    for (var i = 0; i < allTestCases.length; ++i) {
        var expectMatch = expectedIds.indexOf(allTestCases[i].id) >= 0;
        shouldBeEqualToString('getComputedStyle(document.querySelectorAll("#test-root *")[' + i + ']).backgroundColor', expectMatch ? 'rgb(10, 100, 200)' : 'rgb(255, 0, 0)');
    }

    stylingElement.innerHTML = '';
}

function testSelector(selector, expectedIds) {
    debug("Testing \"" + selector + "\"");
    testQuerySelector("#test-root " + selector, expectedIds);
    testStyling("#test-root " + selector, expectedIds);
    debug("");
}

// Test cases inspired by the :nth-child() examples from http://nthmaster.com/.
testSelector(":nth-child(8 of [class], [id])", ["testcase8"]);
testSelector(":nth-child(n+6 of [class], [id])", ["testcase6", "testcase7", "testcase8", "testcase9", "testcase10", "testcase11", "testcase12", "testcase13", "testcase14", "testcase15", "testcase16", "testcase17", "testcase18", "testcase19", "testcase20", "testcase21", "testcase22", "testcase23", "testcase24", "testcase25", "testcase26", "testcase27", "testcase28", "testcase29", "testcase30"]);
testSelector(":nth-child(-n+9 of [class], [id])", ["testcase1", "testcase2", "testcase3", "testcase4", "testcase5", "testcase6", "testcase7", "testcase8", "testcase9"]);
testSelector(":nth-child(n+4 of [class], [id]):nth-child(-n+8 of [class])", ["testcase4", "testcase5", "testcase6", "testcase7", "testcase8"]);
testSelector(":nth-child(-n+8 of [class], [id]):nth-child(n+4 of [class])", ["testcase4", "testcase5", "testcase6", "testcase7", "testcase8"]);
testSelector(":nth-child(n+2 of [class], [id]):nth-child(odd of [class]):nth-child(-n+9 of [class])", ["testcase3", "testcase5", "testcase7", "testcase9"]);
testSelector(":nth-child(3n+1 of [class], [id]):nth-child(even of [class])", ["testcase4", "testcase10", "testcase16", "testcase22", "testcase28"]);
testSelector(":nth-child(even of [class], [id]):nth-child(3n+1 of [class])", ["testcase4", "testcase10", "testcase16", "testcase22", "testcase28"]);

// Test cases inspired by the :nth-of-type() examples from http://nthmaster.com/.
testSelector("testcaseA:nth-child(3 of testcaseA)", ["testcase5"]);
testSelector(":nth-child(3 of testcaseA)", ["testcase5"]);
testSelector("testcaseB:nth-child(4 of testcaseB)", ["testcase8"]);
testSelector(":nth-child(4 of testcaseB)", ["testcase8"]);

testSelector("testcaseA:nth-child(n+3 of testcaseA)", ["testcase5", "testcase7", "testcase9", "testcase11", "testcase13", "testcase15", "testcase17", "testcase19", "testcase21", "testcase23", "testcase25", "testcase27", "testcase29"]);
testSelector(":nth-child(n+3 of testcaseA)", ["testcase5", "testcase7", "testcase9", "testcase11", "testcase13", "testcase15", "testcase17", "testcase19", "testcase21", "testcase23", "testcase25", "testcase27", "testcase29"]);

testSelector("testcaseA:nth-child(2n+2 of testcaseA)", ["testcase3", "testcase7", "testcase11", "testcase15", "testcase19", "testcase23", "testcase27"]);

testSelector("testcaseA:nth-child(-n+4 of testcaseA)", ["testcase1", "testcase3", "testcase5", "testcase7"]);
testSelector("testcaseB:nth-child(-n+5 of testcaseB)", ["testcase2", "testcase4", "testcase6", "testcase8", "testcase10"]);

testSelector("testcaseA:nth-child(n+3 of testcaseA):nth-child(-n+6 of testcaseA)", ["testcase5", "testcase7", "testcase9", "testcase11"]);
testSelector("testcaseA:nth-child(-n+6 of testcaseA):nth-child(n+3 of testcaseA)", ["testcase5", "testcase7", "testcase9", "testcase11"]);

testSelector("testcaseB:nth-child(n+1 of testcaseB):nth-child(-n+3 of testcaseB)", ["testcase2", "testcase4", "testcase6"]);
testSelector("testcaseB:nth-child(-n+3 of testcaseB):nth-child(n+1 of testcaseB)", ["testcase2", "testcase4", "testcase6"]);

testSelector("testcaseA:nth-child(n+3 of testcaseA):nth-child(odd of testcaseA):nth-child(-n+6 of testcaseA)", ["testcase5", "testcase9"]);
testSelector("testcaseB:nth-child(n+1 of testcaseB):nth-child(even of testcaseB):nth-child(-n+3 of testcaseB)", ["testcase4"]);
</script>
<script src="../../resources/js-test-post.js"></script>
</html>
